[libc] Refactor qsort code
This patch makes the following changes:
 - Refactor the internal sorting functions to reduce code duplication.
 - Move the testing machinery done for the testing of `qsort_r` to a
   shared place.

These changes are done in anticipation to the introduction of Annex K's
`qsort_s`. This function shares most of its semantics with `qsort_r`,
therefore most of the testing logic can be shared between the two.
Besides, `qsort`, `qsort_r` and `qsort_r` are all very similar, hence we
can attempt to reduce duplication a bit more.
