+++ libc/src/stdlib/qsort.cpp
@@ -18,11 +18,7 @@ LLVM_LIBC_FUNCTION(void, qsort,
                    (void *array, size_t array_size, size_t elem_size,
                     int (*compare)(const void *, const void *))) {
 
-  const auto is_less = [compare](const void *a, const void *b) -> bool {
-    return compare(a, b) < 0;
-  };
-
-  internal::unstable_sort(array, array_size, elem_size, is_less);
+  internal::unstable_sort(array, array_size, elem_size, compare);
 }
 
 } // namespace LIBC_NAMESPACE_DECL
